requirements.txt
Django>=4.2,<5
dj-database-url
psycopg2-binary
gunicorn
djangorestframework
django-cors-headers
pytest
pytest-django
pytest-cov
whitenoise
